Move ast2 to ast.
