+++ nengo_spa/ast_dynamic.py
@@ -5,7 +5,7 @@ from nengo.network import Network as NengoNetwork
 from nengo.utils.compat import is_number
 import numpy as np
 
-from nengo_spa.ast2 import Fixed, infer_types, Node
+from nengo_spa.ast import Fixed, infer_types, Node
 from nengo_spa.ast_symbolic import FixedScalar, PointerSymbol, Symbol
 from nengo_spa.exceptions import SpaTypeError
 from nengo_spa.types import TAnyVocab, TScalar, TAnyVocabOfDim, TVocabulary
