+++ nengo_spa/tests/test_ast2.py
@@ -4,7 +4,7 @@ from numpy.testing import assert_allclose, assert_equal
 import pytest
 
 import nengo_spa as spa
-from nengo_spa.ast2 import coerce_types
+from nengo_spa.ast import coerce_types
 from nengo_spa.ast_dynamic import ActionSelection
 from nengo_spa.ast_symbolic import PointerSymbol, sym
 from nengo_spa.exceptions import SpaTypeError
